Non-blocking IO development, async

+++ Hazelcast.Net/Hazelcast.Client.Request.Concurrent.Lock/IsLockedRequest.cs
@@ -14,7 +14,7 @@ namespace Hazelcast.Client.Request.Concurrent.Lock
         {
         }
 
-        public IsLockedRequest(Data key, int threadId) : base(key, threadId)
+        public IsLockedRequest(Data key, long threadId) : base(key, threadId)
         {
         }
